Include keys definition in Table case class (#15)
In order to dedup items in batch request, it is necessary to know what are the keys of the table. Changing the signature of Table case class, this allows the keys' definition to be passed around for dedup purpose. This commit also clean up some old IT tests where data is created as resources. Since IT tests are now against dockerised tables, we can use
tables as resources instead.