Build key(s) to AttributeValue internally (#20)
* Build key(s) to AttributeValue internally

Methods used to require Encoder of Key but this can be confusing if key(s) is just a String and value classes. By passing in Table object to these calls, the key(s) is built internally. This is also less error prone.

* Remove Codec for tuple

Codec for tuple can be problematic since it requires both of them to be a Map while they might not be.