Introduce getOpt syntax
Changed `getAs` behaviour to make it fails when the key doesn't exist
rather than default to nul(true). This breaks Decoder[Option[A]] because
it depended on the fact that `getAs` returned `nul(true)` on missing
key. Hence, `getOpt` syntax helps dealing with `nul(true)` and missing
key more explicitly.