Fix series/1.x release pipeline (#66)
Last attempt of releasing from series/1.x branch was bad. The
github action that being used to bump and push tag was not correctly
used (a bump hash tag was used which triggers server bump instead of RC
bump). series/1.x has been deleted and recreated and this commit is to
re-add everything since.